+++ src/shared/bases.js
