+++ tests/shared.test.js
@@ -2,12 +2,12 @@
 
 const WamsShared = require('shared.js');
 
-const reporters = Object.keys(require('shared/Reporters.js'));
+const bases = Object.keys(require('shared/bases.js'));
 const utilities = Object.keys(require('shared/utilities.js'));
 
 const other = ['colours', 'Circle', 'constants', 'IdStamper', 'Message', 'Polygon2D', 'Point2D', 'Rectangle'];
 
-const expected = reporters.concat(utilities).concat(other);
+const expected = bases.concat(utilities).concat(other);
 
 test('Exports all the proper shared utilities', () => {
   const exported = Object.keys(WamsShared);
